server/requirements.txt
grpcio==1.52.0
grpcio-tools==1.52.0